+++ app/controllers/services_controller.rb
@@ -1,95 +1,16 @@
 class ServicesController < ApplicationController
-  include ActionController::Live
-
   before_action :set_service, only: [:show, :edit, :update, :destroy]
+  before_filter :set_sidebar_values, except: [:online_status, :create, :update, :destroy]
+
+  def set_sidebar_values
+    @services = Service.all
+    @plex_services = PlexService.all
+    # @weathers = Weather.all
+  end
 
   # GET /services
   # GET /services.json
   def index
-    tries ||= 3
-    @services = Service.all
-    @plex_services = PlexService.all
-    @weathers = Weather.all
-  rescue ActiveRecord::StatementInvalid => e
-    logger.error "There was an error interacting with the database. The error was: #{e}"
-    sleep(0.25)
-    retry unless (tries -= 1).zero?
-  end
-
-
-  # How to do SSE properly:
-  # https://github.com/rails/rails/blob/6061c540ac7880233a6e32de85cec72c20ed8778/actionpack/lib/action_controller/metal/live.rb#L23
-
-  def notifications
-    response.headers['Content-Type'] = 'text/event-stream'
-    sse = SSE.new(response.stream, retry: 2000)
-    i = 0
-    begin
-      loop do
-        i+= 1
-        events = Array.new
-        @plex_services = PlexService.all
-        @services = Service.all
-        # @weathers = Weather.all
-
-        if @plex_services.empty? && @services.empty?
-          logger.debug 'There were no PlexServices or Generic Services, sleeping for 10s.'
-          sleep(10)
-        end
-
-        @plex_services.try(:each) do |plex_service|
-          plex_service.update_plex_data
-          all_active_sessions = []
-          plex_service.plex_sessions.try(:each) do |plex_session|
-            all_active_sessions << {session_id: plex_session.id,
-                                    html: render_to_string(partial: 'plex_services/now_playing',
-                                                           formats: [:html],
-                                                           locals: {plex_session: plex_session,
-                                                                    active: ''}),
-                                    progress: plex_session.get_percent_done,
-                                    active_sessions: PlexSession.all.ids}
-
-          end
-          if all_active_sessions.length > 0
-            all_active_sessions.each do |active_session|
-              events << {data: active_session, event: 'plex_now_playing'}
-            end
-          elsif i % 5 == 0
-            events << {data: [], event: 'plex_now_playing'}
-          end
-        end
-
-        @services.try(:each) do |service|
-          if i % 5 == 0
-            logger.debug('Looped 5 times, sending all service statuses.')
-            service.ping
-            events << {data: {id: service.id, html: render_to_string(partial: 'service', formats: [:html], locals: {service: service})}.to_json, event: 'online_status'}
-          else
-            unless service.ping_for_status_change.nil?
-              events << {data: {id: service.id, html: render_to_string(partial: 'service', formats: [:html], locals: {service: service})}.to_json, event: 'online_status'}
-            end
-          end
-        end
-
-        # @weathers.try(:each) do |weather|
-        #   weather.get_weather
-        #   events << {data: weather, event: 'weather'}
-        # end
-
-        events.each do |e|
-          sse.write(e[:data], event: e[:event])
-        end
-        sleep 2
-      end
-    rescue IOError
-      logger.warn 'Stream closed: IO Error'
-    rescue ClientDisconnected
-      logger.warn 'Stream closed: Client Disconnect'
-        # rescue StandardError => e
-        #   logger.error "An error occurred during the loop: #{e.message}"
-    ensure
-      sse.close
-    end
   end
 
   # GET /services/:id/online_status
@@ -107,39 +28,26 @@ class ServicesController < ApplicationController
   # GET /services/1
   # GET /services/1.json
   def show
-    @services = Service.all
-    # @service = Service.find(params[:id])
-    @weathers = Weather.all
   end
 
   def all_services
-    @services = Service.all
-    @weathers = Weather.all
   end
 
   # GET /services/new
   def new
-    @services = Service.all
     @service = Service.new
-    @weathers = Weather.all
-    # @plex = Plex.new
   end
 
   def choose_service_type
-    @services = Service.all
-    @weathers = Weather.all
   end
 
   # GET /services/1/edit
   def edit
-    @services = Service.all
-    @weathers = Weather.all
   end
 
   # POST /services
   # POST /services.json
   def create
-    @services = Service.all
     @service = Service.new(service_params)
 
     respond_to do |format|
@@ -156,7 +64,6 @@ class ServicesController < ApplicationController
   # PATCH/PUT /services/1
   # PATCH/PUT /services/1.json
   def update
-    retries ||= 0
     respond_to do |format|
       if @service.update(service_params)
         format.html { redirect_to @service, notice: 'Service was successfully updated.' }
@@ -166,9 +73,6 @@ class ServicesController < ApplicationController
         format.json { render json: @service.errors, status: :unprocessable_entity }
       end
     end
-  rescue ActiveRecord::StatementInvalid
-    logger.warn "Database was probably busy trying to save online status. Trying: #{(retries - 3).abs} more time(s)."
-    retry if (retries += 1) < 3
   end