+++ app/models/service.rb
@@ -13,14 +13,18 @@ class Service < ActiveRecord::Base
   validates_associated :service_flavor
   validates :name, presence: true, uniqueness: true, allow_blank: false
   validates :url, presence: true, uniqueness: true, allow_blank: false
-  validates_inclusion_of :port, in: 1..65535
-  validates :ip, length: {minimum: 7, maximum: 45},
+  validates_inclusion_of :port, in: 1..65_535
+  validates :ip,
+            length: {minimum: 7, maximum: 45},
             format: {with: Resolv::IPv4::Regex},
-            uniqueness: {scope: :port}, allow_blank: true
-  validates :dns_name, length: {minimum: 2, maximum: 127},
-            uniqueness: {scope: :port}, allow_blank: true
-  validates :ip, presence: true, if: (:ip_and_dns_name_dont_exist)
-  validates :dns_name, presence: true, if: (:ip_and_dns_name_dont_exist)
+            uniqueness: {scope: :port},
+            allow_blank: true
+  validates :dns_name,
+            length: {minimum: 2, maximum: 127},
+            uniqueness: {scope: :port},
+            allow_blank: true
+  validates :ip, presence: true, if: :ip_and_dns_name_dont_exist
+  validates :dns_name, presence: true, if: :ip_and_dns_name_dont_exist
 
   def init
     @timeout ||= 5
@@ -29,7 +33,7 @@ class Service < ActiveRecord::Base
 
   def ip_and_dns_name_dont_exist
     if (ip.blank? || ip.to_s.empty?) && (dns_name.blank? || dns_name.to_s.empty?)
-      self.errors.add(:base, 'IP Address or DNS Name must exist')
+      errors.add(:base, 'IP Address or DNS Name must exist')
       true
     else
       false
@@ -44,25 +48,25 @@ class Service < ActiveRecord::Base
   end
 
   def ping
-   ping_status = Rails.cache.fetch("service_#{id}/online", expires_in: 10.seconds) do
+    Rails.cache.fetch("service_#{id}/online", expires_in: 10.seconds) do
       check_online_status
       self.online_status
-   end
-    !ping_status.nil?
+    end
   end
 
   def online_status_string
-    is_online(online_status)
+    online?(online_status)
   end
 
-  def is_online(boolean)
+  def online?(boolean)
     boolean ? 'online' : 'offline'
   end
 
   def ping_for_status_change
     before_ping = self.online_status
-    if before_ping != ping
-      logger.info("Detected status change from #{is_online(before_ping)} to #{online_status_string}")
+    ping_result = ping
+    if before_ping != ping_result
+      logger.info("Detected status change from #{online?(before_ping)} to #{online_status_string} for service: #{name}")
       self.online_status
     else
       nil
@@ -70,10 +74,10 @@ class Service < ActiveRecord::Base
   end
 
   def connect_method
-    if !self.dns_name.blank?
-      self.dns_name
+    if !dns_name.blank?
+      dns_name
     else
-      self.ip
+      ip
     end
   end
 
@@ -82,7 +86,6 @@ class Service < ActiveRecord::Base
     json[:self_uri] = Rails.application.routes.url_helpers.service_online_status_path(id)
     json
   end
-
 
   def online!
     self.online_status = true
@@ -94,6 +97,11 @@ class Service < ActiveRecord::Base
     self
   end
 
+  def last_seen_now!
+    self.last_seen = Time.now
+    self
+  end
+
   def online_status
     Rails.cache.read("service/#{id}/online_status")
   end
@@ -102,25 +110,30 @@ class Service < ActiveRecord::Base
     Rails.cache.write("service/#{id}/online_status", boolean_status)
   end
 
+  def last_seen
+    Rails.cache.read("service/#{id}/last_seen")
+  end
+
+  def last_seen=(timestamp)
+    Rails.cache.write("service/#{id}/last_seen", timestamp)
+  end
+
   private
 
   def check_online_status
-    # retries ||= 0
     ping_destination = connect_method
     begin
       Timeout.timeout(@timeout) do
-        s = TCPSocket.new(ping_destination, self.port)
+        s = TCPSocket.new(ping_destination, port)
         s.close
-        self.online!
-        self.update(last_seen: Time.now)
+        online!
+        last_seen_now!
       end
+        #TODO: Use connection refused to indicate that the server itself is still responding.
     rescue Errno::ECONNREFUSED
-      self.online!
+      offline!
     rescue Timeout::Error, Errno::ENETUNREACH, Errno::EHOSTUNREACH, SocketError
-      self.offline!
+      offline!
     end
-    # rescue ActiveRecord::StatementInvalid
-    #   logger.warn "Database was probably busy trying to save online status. Trying: #{(retries - 3).abs} more time(s)."
-    #   retry if (retries += 1) < 3
   end
 end