+++ config/initializers/lograge.rb
@@ -1,0 +1,10 @@
+ActiveRecord::Base.logger.level = Logger::INFO
+Rails.application.configure do
+  config.lograge.enabled = true
+  config.lograge.custom_options = lambda do |event|
+    exceptions = %w(controller action format id)
+    {
+        params: event.payload[:params].except(*exceptions)
+    }
+  end
+end