+++ config/routes.rb
@@ -1,3 +1,4 @@
+require 'browserlog'
 Rails.application.routes.draw do
 
   root 'services#index'
@@ -6,11 +7,10 @@ Rails.application.routes.draw do
   get 'about' => 'info#about'
   get 'configuration' => 'info#configuration'
   get 'plex_services/all_plex_services'
-
+  get 'notifications' => 'notifications#notifications'
   get 'recently_added' => 'plex_services#recently_added'
 
   resources :services do
-    collection { get :notifications}
     get :online_status
   end
 
@@ -18,4 +18,8 @@ Rails.application.routes.draw do
     get :now_playing
   end
   resources :weather
+
+
+  mount Browserlog::Engine => '/logs'
+
 end