+++ db/migrate/20170329020117_add_stream_type_to_plex_session.rb
@@ -1,0 +1,5 @@
+class AddStreamTypeToPlexSession < ActiveRecord::Migration
+  def change
+    add_column :plex_sessions, :stream_type, :string
+  end
+end