+++ test/controllers/notifications_controller_test.rb
@@ -1,0 +1,14 @@
+require 'test_helper'
+require 'timeout'
+
+class NotificationsControllerTest < ActionController::TestCase
+  # test 'Notifications returns 200 OK' do
+  #   begin
+  #     Timeout.timeout(1) do
+  #       get :notifications
+  #     end
+  #   rescue Timeout::Error
+  #     assert_response :success
+  #   end
+  # end
+end