+++ test/models/service_test.rb
@@ -126,13 +126,15 @@ class ServiceTest < ActiveSupport::TestCase
   # https://en.wikipedia.org/wiki/List_of_TCP_and_UDP_port_numbers
   # Port 9 is discard protocol, so since we can make a handshake with it we know that our online_status checking works.
   test 'ping_for_status_change should be true when service goes online' do
-    service = Service.new(dns_name: '',  ip: '127.0.0.1', port: 9).offline!
-    assert(service.ping_for_status_change, '127.0.0.1:9 should be going from offline to online')
+    service = Service.new(dns_name: '',  ip: '127.0.0.1', port: 6379).offline!
+    Rails.cache.clear
+    assert(service.ping_for_status_change, '127.0.0.1:6379 should be going from offline to online (Status should change)')
   end
 
   test 'ping_for_status_change should be nil when service online_status does not change' do
     service = Service.new(dns_name: '',  ip: '127.0.0.1', port: 9).online!
-    assert_nil(service.ping_for_status_change, '127.0.0.1:9 should be online and be detected as online')
+    Rails.cache.clear
+    assert(!service.ping_for_status_change, '127.0.0.1:9 should be online and be detected as online (No status change)')
   end
 
   # test 'api key must be >= 32 char' do