+++ travisSetup.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+mv example.server_config.yml server_config.yml
 SECRETS="config/secrets.yml"
 SECRET="$(bundle exec rake secret)"
 BOGUS="xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
@@ -9,3 +10,5 @@ echo -e "production:\n secret_key_base: " $SECRET >> $SECRETS
 # rm db/*sqlite3
 bundle exec rake db:create
 bundle exec rake db:migrate
+mkdir -p public/images
+mkdir -p test/test_images